Don't delete authorization credentials after we use them.
Deleting the authorization means that we can't successfully
authenticate against resources in two different domains that use the
same credentials (e.g., www.w3.org and cvs.w3.org).  And deleting it
doesn't really accomplish anything, so we might as well let it stick
around.

This doesn't completely fix the real bug, which is that htmldiff can't
diff resources on different domains with different credentials.  But
that's a problem for another day.

Did some minor style clean-ups while I was in here.